+++ src/Sentry/Laravel/Integration.php
@@ -38,7 +38,7 @@ class Integration implements IntegrationInterface
                 return $event;
             }
 
-            if (null === $event->getTransaction()) {
+            if (empty($event->getTransaction())) {
                 $event->setTransaction($self->getTransaction());
             }
 
@@ -135,7 +135,7 @@ class Integration implements IntegrationInterface
             // Laravel 7 route caching generates a route names if the user didn't specify one
             // theirselfs to optimize route matching. These route names are useless to the
             // developer so if we encounter a generated route name we discard the value
-            if (Str::startsWith($routeName, 'generated::')) {
+            if (Str::contains($routeName, 'generated::')) {
                 $routeName = null;
             }
 
@@ -148,8 +148,16 @@ class Integration implements IntegrationInterface
         }
 
         if (empty($routeName) && $route->getActionName()) {
-            // SomeController@someAction (controller action)
-            $routeName = ltrim($route->getActionName(), (self::$baseControllerNamespace ?? '') . '\\');
+            // Some\Controller@someAction (controller action)
+            $routeName = ltrim($route->getActionName(), '\\');
+
+            $baseNamespace = self::$baseControllerNamespace ?? '';
+
+            // Strip away the base namespace from the action name
+            if (!empty($baseNamespace)) {
+                // @see: Str::after, but this is not available before Laravel 5.4 so we use a inlined version
+                $routeName = array_reverse(explode($baseNamespace . '\\', $routeName, 2))[0];
+            }
         }
 
         if (empty($routeName) || $routeName === 'Closure') {