+++ src/Sentry/Laravel/LogChannel.php
@@ -2,8 +2,9 @@
 
 namespace Sentry\Laravel;
 
-use Illuminate\Log\LogManager;
 use Monolog\Logger;
+use Illuminate\Log\LogManager;
+use Sentry\State\HubInterface;
 
 class LogChannel extends LogManager
 {
@@ -12,10 +13,19 @@ class LogChannel extends LogManager
      *
      * @return Logger
      */
-    public function __invoke(array $config)
+    public function __invoke(array $config): Logger
     {
-        $handler = new SentryHandler($this->app->make('sentry'), $config['level'] ?? Logger::DEBUG, $config['bubble'] ?? true);
+        $handler = new SentryHandler(
+            $this->app->make(HubInterface::class),
+            $config['level'] ?? Logger::DEBUG,
+            $config['bubble'] ?? true
+        );
 
-        return new Logger($this->parseChannel($config), [$this->prepareHandler($handler, $config)]);
+        return new Logger(
+            $this->parseChannel($config),
+            [
+                $this->prepareHandler($handler, $config),
+            ]
+        );
     }
 }