+++ src/Sentry/Laravel/ServiceProvider.php
@@ -2,15 +2,19 @@
 
 namespace Sentry\Laravel;
 
+use Illuminate\Contracts\Http\Kernel as HttpKernelInterface;
+use Illuminate\Foundation\Application as Laravel;
+use Illuminate\Foundation\Http\Kernel as HttpKernel;
+use Illuminate\Log\LogManager;
+use Laravel\Lumen\Application as Lumen;
+use Sentry\ClientBuilder;
+use Sentry\ClientBuilderInterface;
+use Sentry\Integration as SdkIntegration;
+use Sentry\Laravel\Http\LaravelRequestFetcher;
+use Sentry\Laravel\Http\SetRequestIpMiddleware;
 use Sentry\SentrySdk;
 use Sentry\State\Hub;
-use Sentry\ClientBuilder;
 use Sentry\State\HubInterface;
-use Illuminate\Log\LogManager;
-use Sentry\ClientBuilderInterface;
-use Laravel\Lumen\Application as Lumen;
-use Sentry\Integration as SdkIntegration;
-use Illuminate\Foundation\Application as Laravel;
 
 class ServiceProvider extends BaseServiceProvider
 {
@@ -33,10 +37,21 @@ class ServiceProvider extends BaseServiceProvider
      */
     public function boot(): void
     {
-        $this->app->make(static::$abstract);
+        $this->app->make(HubInterface::class);
 
         if ($this->hasDsnSet()) {
             $this->bindEvents($this->app);
+
+            if ($this->app instanceof Lumen) {
+                $this->app->middleware(SetRequestIpMiddleware::class);
+            } elseif ($this->app->bound(HttpKernelInterface::class)) {
+                /** @var \Illuminate\Foundation\Http\Kernel $httpKernel */
+                $httpKernel = $this->app->make(HttpKernelInterface::class);
+
+                if ($httpKernel instanceof HttpKernel) {
+                    $httpKernel->pushMiddleware(SetRequestIpMiddleware::class);
+                }
+            }
         }
 
         if ($this->app->runningInConsole()) {
@@ -56,7 +71,7 @@ class ServiceProvider extends BaseServiceProvider
     public function register(): void
     {
         if ($this->app instanceof Lumen) {
-            $this->app->configure('sentry');
+            $this->app->configure(static::$abstract);
         }
 
         $this->mergeConfigFrom(__DIR__ . '/../../../config/sentry.php', static::$abstract);
@@ -77,7 +92,7 @@ class ServiceProvider extends BaseServiceProvider
     {
         $userConfig = $this->getUserConfig();
 
-        $handler = new EventHandler($this->app->events, $userConfig);
+        $handler = new EventHandler($this->app, $userConfig);
 
         $handler->subscribe();
 
@@ -142,7 +157,7 @@ class ServiceProvider extends BaseServiceProvider
             return $clientBuilder;
         });
 
-        $this->app->singleton(static::$abstract, function () {
+        $this->app->singleton(HubInterface::class, function () {
             /** @var \Sentry\ClientBuilderInterface $clientBuilder */
             $clientBuilder = $this->app->make(ClientBuilderInterface::class);
 
@@ -150,31 +165,40 @@ class ServiceProvider extends BaseServiceProvider
 
             $userIntegrations = $this->resolveIntegrationsFromUserConfig();
 
-            $options->setIntegrations(static function (array $integrations) use ($options, $userIntegrations) {
-                $allIntegrations = array_merge($integrations, $userIntegrations);
+            $options->setIntegrations(function (array $integrations) use ($options, $userIntegrations) {
+                if ($options->hasDefaultIntegrations()) {
+                    // Remove the default error and fatal exception listeners to let Laravel handle those
+                    // itself. These event are still bubbling up through the documented changes in the users
+                    // `ExceptionHandler` of their application or through the log channel integration to Sentry
+                    $integrations = array_filter($integrations, static function (SdkIntegration\IntegrationInterface $integration): bool {
+                        if ($integration instanceof SdkIntegration\ErrorListenerIntegration) {
+                            return false;
+                        }
 
-                if (!$options->hasDefaultIntegrations()) {
-                    return $allIntegrations;
+                        if ($integration instanceof SdkIntegration\ExceptionListenerIntegration) {
+                            return false;
+                        }
+
+                        if ($integration instanceof SdkIntegration\FatalErrorListenerIntegration) {
+                            return false;
+                        }
+
+                        // We also remove the default request integration so it can be readded
+                        // after with a Laravel specific request fetcher. This way we can resolve
+                        // the request from Laravel instead of constructing it from the global state
+                        if ($integration instanceof SdkIntegration\RequestIntegration) {
+                            return false;
+                        }
+
+                        return true;
+                    });
+
+                    $integrations[] = new SdkIntegration\RequestIntegration(
+                        new LaravelRequestFetcher($this->app)
+                    );
                 }
 
-                // Remove the default error and fatal exception listeners to let Laravel handle those
-                // itself. These event are still bubbling up through the documented changes in the users
-                // `ExceptionHandler` of their application or through the log channel integration to Sentry
-                return array_filter($allIntegrations, static function (SdkIntegration\IntegrationInterface $integration): bool {
-                    if ($integration instanceof SdkIntegration\ErrorListenerIntegration) {
-                        return false;
-                    }
-
-                    if ($integration instanceof SdkIntegration\ExceptionListenerIntegration) {
-                        return false;
-                    }
-
-                    if ($integration instanceof SdkIntegration\FatalErrorListenerIntegration) {
-                        return false;
-                    }
-
-                    return true;
-                });
+                return array_merge($integrations, $userIntegrations);
             });
 
             $hub = new Hub($clientBuilder->getClient());
@@ -184,7 +208,7 @@ class ServiceProvider extends BaseServiceProvider
             return $hub;
         });
 
-        $this->app->alias(static::$abstract, HubInterface::class);
+        $this->app->alias(HubInterface::class, static::$abstract);
     }
 
     /**
@@ -194,7 +218,11 @@ class ServiceProvider extends BaseServiceProvider
      */
     private function resolveIntegrationsFromUserConfig(): array
     {
-        $integrations = [new Integration];
+        // Default Sentry Laravel SDK integrations
+        $integrations = [
+            new Integration,
+            new Integration\ExceptionContextIntegration,
+        ];
 
         $userIntegrations = $this->getUserConfig()['integrations'] ?? [];