+++ src/Sentry/Laravel/Tracing/ServiceProvider.php
@@ -5,8 +5,10 @@ namespace Sentry\Laravel\Tracing;
 use Illuminate\Contracts\Http\Kernel as HttpKernelInterface;
 use Illuminate\Contracts\View\Engine;
 use Illuminate\Contracts\View\View;
+use Illuminate\Foundation\Http\Kernel as HttpKernel;
 use Illuminate\View\Engines\EngineResolver;
 use Illuminate\View\Factory as ViewFactory;
+use Laravel\Lumen\Application as Lumen;
 use Sentry\Laravel\BaseServiceProvider;
 
 class ServiceProvider extends BaseServiceProvider
@@ -18,11 +20,15 @@ class ServiceProvider extends BaseServiceProvider
 
             $this->bindViewEngine();
 
-            if ($this->app->bound(HttpKernelInterface::class)) {
-                /** @var \Illuminate\Contracts\Http\Kernel $httpKernel */
+            if ($this->app instanceof Lumen) {
+                $this->app->middleware(Middleware::class);
+            } elseif ($this->app->bound(HttpKernelInterface::class)) {
+                /** @var \Illuminate\Foundation\Http\Kernel $httpKernel */
                 $httpKernel = $this->app->make(HttpKernelInterface::class);
 
-                $httpKernel->prependMiddleware(Middleware::class);
+                if ($httpKernel instanceof HttpKernel) {
+                    $httpKernel->prependMiddleware(Middleware::class);
+                }
             }
         }
     }
@@ -69,7 +75,7 @@ class ServiceProvider extends BaseServiceProvider
         /** @var ViewFactory $viewFactory */
         $viewFactory = $this->app->make('view');
 
-        $viewFactory->composer('*', static function (View $view) use ($viewFactory) : void {
+        $viewFactory->composer('*', static function (View $view) use ($viewFactory): void {
             $viewFactory->share(ViewEngineDecorator::SHARED_KEY, $view->name());
         });